🔨Clusters maintenance script: refactoring and add SSH tunneling (#5886)
