+++ scripts/maintenance/computational-clusters/Makefile
@@ -2,14 +2,24 @@
 
 SHELL := /bin/bash
 
-install:
+.venv:
 	# creating python virtual environment
 	@uv venv .venv
-	# activating python virtual environment
-	@source .venv/bin/activate
 	# installing python dependencies
 	@uv pip install --upgrade pip setuptools wheel
-	@uv pip install -r requirements.txt
+
+
+install: .venv
+	# activating python virtual environment
+	@source .venv/bin/activate
+	# installing package
+	@uv pip install .
 	# now you can call the maintenance scripts
 	# source .venv/bin/activate
-	# e.g. ./osparc_clusters.py PATH/TO/REPO.CONFIG --ssh-key-path=PATH/TO/SSHKEY
+	# autoscaled-monitor --deploy-config PATH/TO/REPO.CONFIG summary
+
+install-dev: .venv
+	# activating python virtual environment
+	@source .venv/bin/activate
+	# installing package
+	@uv pip install -e .
