+++ scripts/maintenance/computational-clusters/pyproject.toml
@@ -1,0 +1,37 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+dependencies = [
+  "arrow",
+  "aiocache",
+  "asyncpg",
+  "black",
+  "boto3",
+  # NOTE: these must be in sync with ospar
+  "cloudpickle",
+  "dask[distributed]",
+  "mypy_boto3_ec2",
+  "types-boto3",
+  "parse",
+  "paramiko",
+  "pydantic[email]",
+  "pylint",
+  "python-dotenv",
+  "typer",
+  "rich",
+  "sqlalchemy[asyncio]",
+  "sshtunnel",
+]
+name = "autoscaled-monitor"
+version = "1.0.0"
+authors = [
+  { name = "Sylvain Anderegg", email = "35365065+sanderegg@users.noreply.github.com" },
+]
+description = "Helper script for monitoring clusters"
+readme = "README.md"
+requires-python = ">=3.10"
+
+[project.scripts]
+autoscaled-monitor = "autoscaled_monitor.cli:app"
